Initial commit
Backend for tabular-browser. Implement API endpoints for querying and inserting data.
The server uses a single API endpoint named /query. The endpoint expects a JSON body
defining the query to be performed, query can be of either read or write type.
JSON body is parsed to generate SQL statement to be executed against the DB.
Errorsi in case of query execution are handled and returned to the client for further propogation.